+++ GRDB/FTS/FTS5.swift
@@ -32,7 +32,7 @@
         }
         
         /// Don't use this method.
-        public func moduleArguments(_ definition: FTS5TableDefinition) -> [String] {
+        public func moduleArguments(for definition: FTS5TableDefinition, in db: Database) throws -> [String] {
             var arguments: [String] = []
             
             if definition.columns.isEmpty {
@@ -51,13 +51,21 @@
                 arguments.append("tokenize=\(tokenizer.components.joined(separator: " ").sqlExpression.sql)")
             }
             
-            if let content = definition.content {
-                arguments.append("content=\(content.sqlExpression.sql)")
+            switch definition.contentMode {
+            case .raw(let content, let contentRowID):
+                if let content = content {
+                    arguments.append("content=\(content.sqlExpression.sql)")
+                }
+                if let contentRowID = contentRowID {
+                    arguments.append("content_rowid=\(contentRowID.sqlExpression.sql)")
+                }
+            case .synchronized(let contentTable):
+                arguments.append("content=\(contentTable.sqlExpression.sql)")
+                if let rowIDColumn = try db.primaryKey(contentTable)?.rowIDColumn {
+                    arguments.append("content_rowid=\(rowIDColumn.sqlExpression.sql)")
+                }
             }
             
-            if let contentRowID = definition.contentRowID {
-                arguments.append("content_rowid=\(contentRowID.sqlExpression.sql)")
-            }
             
             if let prefixes = definition.prefixes {
                 arguments.append("prefix=\(prefixes.map { "\($0)" }.joined(separator: " ").sqlExpression.sql)")
@@ -72,6 +80,55 @@
             }
             
             return arguments
+        }
+        
+        /// Reserved; part of the VirtualTableModule protocol.
+        ///
+        /// See Database.create(virtualTable:using:)
+        public func database(_ db: Database, didCreate tableName: String, using definition: FTS5TableDefinition) throws {
+            switch definition.contentMode {
+            case .raw:
+                break
+            case .synchronized(let contentTable):
+                // https://sqlite.org/fts5.html#external_content_tables
+                
+                let rowIDColumn = (try? db.primaryKey(contentTable))??.rowIDColumn ?? "rowid"
+                let ftsTable = tableName.quotedDatabaseIdentifier
+                let content = contentTable.quotedDatabaseIdentifier
+                let indexedColumns = definition.columns.map { $0.name }
+                
+                let ftsColumns = (["rowid"] + indexedColumns)
+                    .map { $0.quotedDatabaseIdentifier }
+                    .joined(separator: ", ")
+                
+                let newContentColumns = ([rowIDColumn] + indexedColumns)
+                    .map { "new.\($0.quotedDatabaseIdentifier)" }
+                    .joined(separator: ", ")
+                
+                let oldContentColumns = ([rowIDColumn] + indexedColumns)
+                    .map { "old.\($0.quotedDatabaseIdentifier)" }
+                    .joined(separator: ", ")
+                
+                try db.execute(
+                    "CREATE TRIGGER \("__\(contentTable)_ai".quotedDatabaseIdentifier) AFTER INSERT ON \(content) BEGIN " +
+                        "INSERT INTO \(ftsTable)(\(ftsColumns)) VALUES (\(newContentColumns)); " +
+                    "END")
+                
+                try db.execute(
+                    "CREATE TRIGGER \("__\(contentTable)_ad".quotedDatabaseIdentifier) AFTER DELETE ON \(content) BEGIN " +
+                        "INSERT INTO \(ftsTable)(\(ftsTable), \(ftsColumns)) VALUES('delete', \(oldContentColumns)); " +
+                    "END")
+                
+                try db.execute(
+                    "CREATE TRIGGER \("__\(contentTable)_au".quotedDatabaseIdentifier) AFTER UPDATE ON \(content) BEGIN " +
+                        "INSERT INTO \(ftsTable)(\(ftsTable), \(ftsColumns)) VALUES('delete', \(oldContentColumns)); " +
+                        "INSERT INTO \(ftsTable)(\(ftsColumns)) VALUES (\(newContentColumns)); " +
+                    "END")
+                
+                // https://sqlite.org/fts5.html#the_rebuild_command
+                
+                try db.execute("INSERT INTO \(ftsTable)(\(ftsTable)) VALUES('rebuild')")
+            }
         }
         
         static func api(_ db: Database) -> UnsafePointer<fts5_api>? {
@@ -103,7 +160,13 @@
     ///
     /// See https://www.sqlite.org/fts5.html
     public final class FTS5TableDefinition {
+        enum ContentMode {
+            case raw(content: String?, contentRowID: String?)
+            case synchronized(String)
+        }
+        
         fileprivate var columns: [FTS5ColumnDefinition] = []
+        fileprivate var contentMode: ContentMode = .raw(content: nil, contentRowID: nil)
         
         /// The virtual table tokenizer
         ///
@@ -116,13 +179,73 @@
         
         /// The FTS5 `content` option
         ///
+        /// When you want the full-text table to be synchronized with the
+        /// content of an external table, prefer the `synchronize(withTable:)`
+        /// method.
+        ///
+        /// Setting this property invalidates any synchronization previously
+        /// established with the `synchronize(withTable:)` method.
+        ///
         /// See https://www.sqlite.org/fts5.html#external_content_and_contentless_tables
-        public var content: String?
+        public var content: String? {
+            get {
+                switch contentMode {
+                case .raw(let content, _):
+                    return content
+                case .synchronized(let contentTable):
+                    return contentTable
+                }
+            }
+            set {
+                switch contentMode {
+                case .raw(_, let contentRowID):
+                    contentMode = .raw(content: newValue, contentRowID: contentRowID)
+                case .synchronized:
+                    contentMode = .raw(content: newValue, contentRowID: nil)
+                }
+            }
+        }
         
         /// The FTS5 `content_rowid` option
         ///
+        /// When you want the full-text table to be synchronized with the
+        /// content of an external table, prefer the `synchronize(withTable:)`
+        /// method.
+        ///
+        /// Setting this property invalidates any synchronization previously
+        /// established with the `synchronize(withTable:)` method.
+        ///
         /// See https://sqlite.org/fts5.html#external_content_tables
-        public var contentRowID: String?
+        public var contentRowID: String? {
+            get {
+                switch contentMode {
+                case .raw(_, let contentRowID):
+                    return contentRowID
+                case .synchronized:
+                    return nil
+                }
+            }
+            set {
+                switch contentMode {
+                case .raw(let content, _):
+                    contentMode = .raw(content: content, contentRowID: newValue)
+                case .synchronized:
+                    contentMode = .raw(content: nil, contentRowID: newValue)
+                }
+            }
+        }
+        
+        /// Synchronizes the full-text table with the content of an external
+        /// table.
+        ///
+        /// The full-text table is initially populated with the existing
+        /// content in the external table. SQL triggers make sure that the
+        /// full-text table is kept up to date with the external table.
+        ///
+        /// See https://sqlite.org/fts5.html#external_content_tables
+        public func synchronize(withTable tableName: String) {
+            contentMode = .synchronized(tableName)
+        }
         
         /// Support for the FTS5 `prefix` option
         ///